+++ OmmelSamvirke.Application.UnitTests/Features/Pages/PageTemplates/Commands/Common/PageTemplateCommandsTestBase.cs
@@ -1,6 +1,7 @@
 ﻿using OmmelSamvirke.Application.Features.Pages.PageTemplates.DTOs;
 using OmmelSamvirke.Domain.Features.Pages.Enums;
 using OmmelSamvirke.Domain.Features.Pages.Models;
+using OmmelSamvirke.Domain.Features.Pages.Models.ContentBlockData;
 using OmmelSamvirke.Domain.Features.Pages.Models.ContentBlocks;
 
 namespace OmmelSamvirke.Application.UnitTests.Features.Pages.PageTemplates.Commands.Common;
@@ -9,10 +10,12 @@ public abstract class PageTemplateCommandsTestBase
 {
     protected Mock<IPageTemplateRepository> PageTemplateRepository = null!;
     protected Mock<IPageRepository> PageRepository = null!;
+    protected Mock<IContentBlockDataRepository> ContentBlockDataRepository = null!;
     protected Mock<IMapper> Mapper = null!;
     protected PageTemplate DefaultPageTemplate = null!;
     protected PageTemplate DefaultArchivedPageTemplate = null!;
     protected PageTemplateDto DefaultPageTemplateDto = null!;
+    protected Page DefaultPage = null!;
     protected const int DefaultPageTemplateId = 1;
 
     [SetUp]
@@ -21,6 +24,7 @@ public abstract class PageTemplateCommandsTestBase
         Mapper = new Mock<IMapper>();
         PageTemplateRepository = new Mock<IPageTemplateRepository>();
         PageRepository = new Mock<IPageRepository>();
+        ContentBlockDataRepository = new Mock<IContentBlockDataRepository>();
         DefaultPageTemplate = new PageTemplate(
             DefaultPageTemplateId,
             DateTime.Now, 
@@ -42,6 +46,13 @@ public abstract class PageTemplateCommandsTestBase
             new HashSet<Layouts>(),
             new List<ContentBlock>(),
             PageTemplateState.Public
+        );
+        DefaultPage = new Page(
+            1,
+            DateTime.Now,
+            DateTime.Now,
+            "TestPage",
+            DefaultPageTemplate
         );
     }
 }