+++ OmmelSamvirke.Domain/Features/Pages/Interfaces/ContentBlockData/IContentBlockData.cs
@@ -1,0 +1,9 @@
+﻿using OmmelSamvirke.Domain.Features.Pages.Models.ContentBlocks;
+
+namespace OmmelSamvirke.Domain.Features.Pages.Interfaces.ContentBlockData;
+
+public interface IContentBlockData
+{
+    ContentBlock BaseContentBlock { get; }
+    int PageId { get; }
+}