+++ OmmelSamvirke.Domain/Features/Pages/Interfaces/Repositories/IContentBlockDataRepository.cs
@@ -1,0 +1,8 @@
+﻿using OmmelSamvirke.Domain.Features.Pages.Interfaces.ContentBlockData;
+
+namespace OmmelSamvirke.Domain.Features.Pages.Interfaces.Repositories;
+
+public interface IContentBlockDataRepository
+{
+    Task<List<IContentBlockData>> GetByPageIdAsync(int pageId);
+}