+++ OmmelSamvirke.Domain/changelog.md
@@ -5,6 +5,11 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [0.8.0] - 2023-07-03
+### Changed
+- Added interface IContentBlockData which can be used when a collection
+of ContentBlockData elements is needed.
+
 ## [0.7.0] - 2023-07-01
 ### Added
 - Domain classes for the Pages feature